Added some missile stuff  to the main
Added the player ship, messed around in the player class by adding a default constructor and properties for rectangle and texture for drawing the ship.  Wasn't sure on how the shot delay worked, so I didn't implement that. The ship moves left and right and shoots a missile when space is pressed.